requirements.txt
torch=2.0.0
gymnasium=0.28.1
matplotlib=3.7.1
pygame=2.4.0
networkx=3.1
transformers=4.29.2
stable-baselines3=2.2.0a0
sb3-contrib=2.1.0
wandb=0.15.1
hydra-core=1.3.2
tensorboard